requirements.txt
boto3>=1.35.0,<2.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
lxml>=5.1.0,<6.0.0
urllib3>=2.0.0,<3.0.0
pyspark==3.5.2
pandas>=2.1.0
pyarrow>=14.0.0,<17.0.0
python-dotenv>=1.0.0